Changed BPM calculation to float
BPM Calculation was not correct for multiple reasons. Fixed calculation and changed data type to float to allow for more accurate spawning.

+++ Assets/Scripts/GameManager.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 public class GameManager : MonoBehaviour
 {
     public AudioSource gameSong;
-    public int songBPM = 110;
+    public float songBPM = 110.0f;
     public bool godMode = false;
     [SerializeField]
     private int health = 3;
